code reorganised
