+++ index.html
@@ -1,61 +1,94 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>To Do</title>
-    <link rel="stylesheet" href="styles.css">
-    <link rel="stylesheet" href="styles-responsive.css">
-    <script src="script.js" type="module"></script>
-</head>
-<body>
+    
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    <link rel="stylesheet" type="text/css" href="./css/styles.css" />
+    <link rel="stylesheet" type="text/css" href="./css/styles-responsive.css" />
+    <script src="./javascript/script.js" type="module"></script>
+    <script src="./javascript/modal-operations.js" type="module"></script>
+  </head>
+  <body>
     <div class="title_div">
-        <a href="#"><img src="./images/icons8-notepad-64.png"></a>
-        <a href="#" class="titlelink"><h1 class="heading">To Do</h1></a>
+      <a href="#"><img src="./images/icons8-notepad-64.png" /></a>
+      <a href="#" class="titlelink"><h1 class="heading">To Do</h1></a>
     </div>
-    <div class="searchBar" >
-        <form id = "searchArea" class="searchArea">
-            <input name="searchText" type="text" placeholder="Search.." id="searchText">
-            <img src="./images/x.svg" id="clear-search" >
-        </form>
+    <div class="searchBar">
+      <form id="searchArea" class="searchArea">
+        <input
+          name="searchText"
+          type="text"
+          placeholder="Search.."
+          id="searchText"
+        />
+        <img src="./images/x.svg" id="clear-search" />
+      </form>
     </div>
     <div class="todo_list">
-        <div  class="list-items" id="list-items">
-            <ol id="printing-list"> 
+      <div class="list-items" id="list-items">
+        <ol id="printing-list"></ol>
+      </div>
+      <div id="floatDiv" class=""></div>
+      <a class="floatButton" id="floatButton">
+        <img class="plus" id="plus" src="./images/plus.svg" />
+      </a>
+      <form id="input-area" class="input-area">
+        <input
+          name="new-note-input"
+          type="text"
+          placeholder="Enter task.."
+          id="new-note-input"
+        />
+        <input type="submit" class="submission" id="button" value="" />
+      </form>
+    </div>
 
-            </ol>
-        </div>
-        <div id="floatDiv" class="">
-        </div>
-        <a class="floatButton" id="floatButton">
-            <img class="plus" id="plus" src="./images/plus.svg">
-        </a>
-        <form id = "input-area" class="input-area">
-            <input name="new-note-input" type="text" placeholder="Enter task.." id="new-note-input">
-            <input type="submit" class="submission" id="button" value="">
-        </form>
+    <div class="overlay" id="overlay">
+      <button id="mobile-edit-button" class="mobile-edit-button">Edit</button>
+      <button id="mobile-delete-button" class="mobile-delete-button">
+        Delete
+      </button>
+      <button
+        id="bottom-drawer-cancel-button"
+        class="bottom-drawer-cancel-button"
+      >
+        Cancel
+      </button>
     </div>
-    <div class= 'modalClose' id="emptyModalParent">
-        <div class= "modal" id="modal">
-            <div class=emptyModalContent>
-                <div class="headingInsideModal">Nothing Entered</div>
-                <div class='contentInsideModal'>Enter a description to add to the task list</div>
-            </div>
-            <button class='emptyModalButton normalButton' onClick="window.closeModal();">Okay</button>
+
+    <div class="hidden" id="empty-input-modal">
+      <div class="modal" id="empty-input-modal-content">
+        <div class="emptyModalContent">
+          <div class="headingInsideModal">Nothing Entered</div>
+          <div class="contentInsideModal">
+            Enter a description to add to the task list
+          </div>
         </div>
+        <button class="emptyModalButton normalButton" id="close-empty-input-modal-button">Okay</button>
+      </div>
     </div>
-    <div class= 'modalClose' id="removeModal">
-        <div class= "modal" id="popup">
-            <div>
-                <div class="headingInsideModal">Confirm Deletion?</div>
-                <div class='contentInsideModal'>Are you sure to permenently remove the selected item?</div>
-            </div>
-            <div class='removeModalButtons'>
-                <button class='normalButton' id= 'close'>Cancel</button>
-                <button class='dangerButton' id= 'confirm'>Delete</button>
-            </div>
+
+    <div class="hidden modal-backdrop" id="deletion-modal">
+      <div class="modal" id="deletion-modal-content">
+        <div>
+          <div class="headingInsideModal">Confirm Deletion?</div>
+          <div class="contentInsideModal">
+            Are you sure to permenently remove the selected item?
+          </div>
         </div>
+        <div class="deletionModalButtons">
+          <button class="normalButton" id="close-deletion-modal-button">
+            Cancel
+          </button>
+          <button class="dangerButton" id="confirm-deletion-button">
+            Delete
+          </button>
+        </div>
+      </div>
     </div>
-</body>
+  </body>
 </html>
