+++ javascript/util.js
@@ -1,0 +1,19 @@
+const placeCaretAtEnd = (element) => {
+    element.focus();
+    if (window.getSelection != "undefined"
+        && document.createRange != "undefined") {
+        var range = document.createRange();
+        range.selectNodeContents(element);
+        range.collapse(false);
+        var sel = window.getSelection();
+        sel.removeAllRanges();
+        sel.addRange(range);
+    } else if (document.body.createTextRange != "undefined") {
+        var textRange = document.body.createTextRange();
+        textRange.moveToElementText(element);
+        textRange.collapse(false);
+        textRange.select();
+    }
+}
+
+export { placeCaretAtEnd };
