--- styles-responsive.css
@@ -1,186 +1,0 @@
-@media (min-width: 700px){
-    .list-items {
-      height: calc(100vh - 185px);
-    }
-    .hide{
-      display:none;
-    }
-  }
-  @media (max-width:700px){
-    .floatDiv{
-      position: absolute;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      top: 0;
-      height: 100vh;
-      width: 100vw;
-      background-color: #00000088;
-    }
-    .searchBar{
-      position: fixed;
-      top: 55px;
-      width: 100%;
-    }
-    .searchArea{
-      width: inherit;
-      margin:0px;
-    }
-    #searchText{
-      border:none;
-      padding: 13px 20px 13px 20px;
-      background-position: 12px 10px;
-      text-indent: 20px;
-    }
-    .todo_list{
-      align-items:stretch;
-      margin-top: 42px;
-      width: 100%;
-    }
-    .list-content{
-      margin-left: 1rem;
-    }
-    .checkbox{
-      margin-right: 14px;
-    }
-    .floatButton{
-      position: fixed;
-      border-radius: 20px;
-      height: 40px;
-      width: 40px;
-      bottom:20px;
-      right:20px;
-      border: 0.5px solid black;
-      box-shadow: 0 3px 3px 0 rgb(139 139 139);
-      background-color: #220083;
-      color:#FFF;
-      text-align:center;
-      box-shadow: 2px 2px 3px #999;
-    }
-    .plus{
-      margin-top: 10px;
-    }
-    .input-area{
-      display:none;
-      bottom: 0px;
-    }
-    #new-note-input{
-      margin:0px;
-      border-radius: 0px;
-    }
-    .submission{
-      position: absolute;
-      right: 0px;
-      height: 30px;
-      width: 30px;
-      border-radius: 15px;
-      background-position: 40% 67%;
-      margin-right: 10px;
-      background-size: auto 20px;
-    }
-    
-    .dropdown{
-      display: flex;
-      margin-right: 14px
-    }
-    .dotmenu{
-      width:18px;
-      display: flex;
-      align-items: center;
-    }
-    .buttons{
-      display: none;
-    }
-    .overlay{
-        position: fixed;
-        bottom:0px;
-        left: 10%;
-        right: 10%;
-        display:none;
-        flex-direction: column;
-        width: 80%;
-        padding-top: 20px;
-        padding-bottom: 20px;
-        background-color: #FFF;
-        border-radius: 13px 13px 0px 0px;
-        z-index: 1;
-    }
-    .overlay button{
-        padding: 10px;
-        font-family: inherit;
-        background-color: #FFF;
-        border-top: none;
-        border-right: none;
-        border-left: none;
-        border-bottom: 1px solid rgb(212, 212, 212);
-    }
-    .overlay button:hover{
-        cursor: pointer;
-        background-color:rgba(0, 0, 0, 0.1);
-    }
-    .modal{
-        width: 300px;
-        height: 145px;
-    }
-    .emptyModalParent{
-        height: 100vh;
-        width: 100vw;
-    }
-    .emptyModalButton{
-      top: 0px;
-      left: 169px;
-    }
-    .emptyModalContent{
-      padding-left: 5px;
-    }
-    .headingInsideModal{
-      font-size: large;
-      font-weight: 400;
-    }
-    .contentInsideModal{
-      font-size: medium;
-    }
-    .normalButton{
-      min-height: 29px;
-      min-width: 73px;
-      border-radius: 13px;
-    }
-    .removeModalButtons{
-        top: 0.5rem;
-        left: 6rem;
-    }
-    .dangerButton{
-        min-height: 29px;
-        min-width: 73px;
-        border-radius: 13px;
-    }
-    .add-empty-image{
-        margin-top: 40%;
-    }
-    .saveButton{
-      position: absolute;
-      right: 18px;
-      border-radius: 15px;
-      height: 30px;
-      width: 30px;
-      background-color: #220083;
-      outline: none;
-      border: 0.5px solid black;
-      margin: 0 1rem;
-      box-shadow: 0 3px 3px 0 rgb(139, 139, 139);
-    }
-    .saveImage{
-      position: relative;
-      right: 0px;
-      max-width: 25px;
-    }
-    .empty {
-      max-height: 8rem;
-      max-width: 8rem;
-      color: #00488a;
-    }
-  }
-  .flexDisplay{
-    display:flex;
-  }
-
